backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
Django==4.0.5
django-cors-headers==3.12.0
django-filter==22.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
